+++ src/test/scala/com/dataintuitive/viash/MainTestDockerSuite.scala
@@ -242,7 +242,7 @@ class MainTestDockerSuite extends FunSuite with BeforeAndAfterAll {
       configFile
     )
 
-    assert(testText.contains("anndata")) // check whether custom package was picked up
+    assert(testText.contains("custom_target_image_tag")) // check whether custom package was picked up
     assert(testText.contains("Running tests in temporary directory: "))
     assert(testText.contains("SUCCESS! All 2 out of 2 test scripts succeeded!"))
     assert(testText.contains("Cleaning up temporary directory"))