+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# Viash 0.5.6
+
+## BREAKING CHANGES
+
+* `BashWrapper`: Forbidden flags `-v`, `--verbose`, `--verbosity` have been renamed to `---v`, `---verbose`, `---verbosity`.
+
+## MINOR CHANGES
+
+* Set version of helper scripts to the same version as Viash.
+
+* `DockerPlatform`: Produce helpful warning message when Docker image can't be found remotely (#94).
+
+* `DockerPlatform`: Produce helpful error message when Docker isn't installed or the daemon is not running (#94 bis).
+
+## BUG FIXES
+
+* `viash_install`:
+  - Passing Viash path as a string instead of as a file to ensure the path is not converted to an absolute path
+  - Switch from Docker backend to a Native backend, 'unzip' and 'wget' are required.
+  - Correctly set the log file for viash_test.
+  
+* `DockerPlatform`: Added sleep workaround to avoid concurrency issue where a file is executed to
+  build docker containers but apparently still in the process of being written.
+  
+* `DockerPlatform`: Fix order issue of ---verbose flag in combination with ---setup, allowing to run 
+  `viash run config.vsh.yaml -- ---setup cb ---verbose` and actually get output.
+  
+
 # Viash 0.5.5
 
 ## BREAKING CHANGES